+++ Code/wwlib/chunkio.h
@@ -254,12 +254,12 @@ private:
 */
 #define WRITE_WWSTRING_CHUNK(csave,id,var) { \
 	csave.Begin_Chunk(id); \
-	csave.Write((const char *)var, var.Get_Length () + 1); \
+	csave.Write((const char *)var, static_cast<int>(var.Get_Length () + 1)); \
 	csave.End_Chunk(); }
 
 #define WRITE_WIDESTRING_CHUNK(csave,id,var) { \
 	csave.Begin_Chunk(id); \
-	csave.Write((const wchar_t *)var, (var.Get_Length () + 1) * 2); \
+	csave.Write((const wchar_t *)var, static_cast<int>((var.Get_Length () + 1) * 2)); \
 	csave.End_Chunk(); }
 
 
@@ -279,10 +279,10 @@ private:
 **
 */
 #define READ_WWSTRING_CHUNK(cload,id,var)		\
-	case (id):	cload.Read(var.Get_Buffer(cload.Cur_Chunk_Length()),cload.Cur_Chunk_Length()); break;	\
+	case (id):	cload.Read(var.Get_Buffer(static_cast<size_t>(cload.Cur_Chunk_Length())),cload.Cur_Chunk_Length()); break;	\
 
 #define READ_WIDESTRING_CHUNK(cload,id,var)		\
-	case (id):	cload.Read(var.Get_Buffer((cload.Cur_Chunk_Length()+1)/2),cload.Cur_Chunk_Length()); break;	\
+	case (id):	cload.Read(var.Get_Buffer(static_cast<size_t>((cload.Cur_Chunk_Length()+1)/2)),cload.Cur_Chunk_Length()); break;	\
 
 
 /*
@@ -323,12 +323,12 @@ private:
 
 #define WRITE_MICRO_CHUNK_WWSTRING(csave,id,var) { \
 	csave.Begin_Micro_Chunk(id); \
-	csave.Write((const char *)var, var.Get_Length () + 1); \
+	csave.Write((const char *)var, static_cast<int>(var.Get_Length () + 1)); \
 	csave.End_Micro_Chunk(); }
 
 #define WRITE_MICRO_CHUNK_WIDESTRING(csave,id,var) { \
 	csave.Begin_Micro_Chunk(id); \
-	csave.Write((const wchar_t *)var, (var.Get_Length () + 1) * 2); \
+	csave.Write((const wchar_t *)var, static_cast<int>((var.Get_Length () + 1) * 2)); \
 	csave.End_Micro_Chunk(); }
